02 - Recursos Humanos (RH)/Data_Talent_Scout_AI/app/backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-dotenv
tqdm
pandas
pyarrow
boto3
asyncpg
pydantic
crewai
crewai[tools]